+++ src/main/java/com/klarna/hiverunner/HiveRunnerExtension.java
@@ -37,7 +37,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.base.Preconditions;
-import com.google.common.base.Predicates;
 import com.klarna.hiverunner.annotations.HiveRunnerSetup;
 import com.klarna.hiverunner.annotations.HiveSQL;
 import com.klarna.hiverunner.builder.Script;
@@ -74,15 +73,9 @@ public class HiveRunnerExtension implements AfterEachCallback, TestInstancePostP
     scriptsUnderTest = container.getScriptsUnderTest();
   }
 
-    private void setupConfig(Object target) {
-        final Set<Field> fields = ReflectionUtils
-                .getAllFields(
-                        target.getClass(),
-                        withType(HiveRunnerConfig.class)
-                                .and(
-                                    withAnnotation(HiveRunnerSetup.class)
-                                )
-                );
+  private void setupConfig(Object target) {
+    Set<Field> fields = ReflectionUtils.getAllFields(target.getClass(),
+            withType(HiveRunnerConfig.class).and(withAnnotation(HiveRunnerSetup.class)));
 
     Preconditions.checkState(fields.size() <= 1,
         "Only one field of type HiveRunnerConfig should be annotated with @HiveRunnerSetup");