+++ src/test/java/com/klarna/hiverunner/ToUpperCaseSerDe.java
@@ -27,7 +27,11 @@ import org.apache.hadoop.hive.serde2.objectinspector.primitive.PrimitiveObjectIn
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.Writable;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Objects;
+import java.util.Properties;
 
 public class ToUpperCaseSerDe extends AbstractSerDe {